fix: make flowWarnings show to unlock saving
closes #8115
